feat: update @supabase/supabase-js dependency to version 2.50.0; refactor context imports in components and implement resource fetching from Supabase
