+++ package.json
@@ -10,7 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@supabase/supabase-js": "^2.39.0",
+    "@supabase/supabase-js": "^2.50.0",
     "brace-expansion": "^2.0.2",
     "framer-motion": "^10.16.16",
     "lucide-react": "^0.344.0",
